User mesages are now posted to the server
But I had to copy a function to recall the messages that was posted to
the server because you can't access that function within a userEffect
hook.

Furthermore, there is a warhing that newMessage is assinged but never
used, and this si because in order to get Axios to post the message it
has to be stored somewhere.  So I am stuck with that warning.